+++ src/common/settings.ts
@@ -1,6 +1,6 @@
 //spell-checker:words vscroll, hscroll, askoncancel, askonclose,commandon, cmdfont
 //spell-checker:ignore emoteto, emotetos askonchildren YYYYMMDD Hmmss
-import { NewLineType, Log, BackupSelection, TrayClick, OnDisconnect, ProfileSortOrder } from './types';
+import { NewLineType, Log, BackupSelection, TrayClick, OnDisconnect, ProfileSortOrder, OnProfileChange } from './types';
 const path = require('path');
 const fs = require('fs');
 
@@ -453,6 +453,8 @@ export class Settings {
 
     public allowEval: boolean = true;
     public externalWho: boolean = true;
+    public watchForProfilesChanges = false;
+    public onProfileChange: OnProfileChange = OnProfileChange.Nothing;
 
     public windows = {};
     public buttons = {